:hammer: refactor upload file and reset password token

+++ src/modules/permissions/permissions.service.js.js
@@ -6,6 +6,11 @@ import { calculatePagination } from "#src/utils/pagination.util";
 const SELECTED_FIELDS =
   "_id name description module endpoint method status";
 
+/**
+ * Create permission
+ * @param {*} data
+ * @returns
+ */
 export async function createPermissionService(data) {
   const { name } = data;
   const isExistName = await checkExistPermissionNameService(name);
@@ -17,6 +22,12 @@ export async function createPermissionService(data) {
   return await findPermissionByIdService(permission._id);
 }
 
+/**
+ * Find all permissions
+ * @param {*} query
+ * @param {*} selectFields
+ * @returns
+ */
 export async function findAllPermissionsService(
   query,
   selectFields = SELECTED_FIELDS
@@ -39,7 +50,8 @@ export async function findAllPermissionsService(
   const permissions = await PermissionModel.find(filterOptions)
     .skip(metaData.offset)
     .limit(metaData.limit)
-    .select(selectFields);
+    .select(selectFields)
+    .sort({ createdAt: -1 });
 
   return {
     list: permissions,
@@ -47,11 +59,16 @@ export async function findAllPermissionsService(
   };
 }
 
+/**
+ * Find permission by id
+ * @param {*} id
+ * @param {*} selectFields
+ * @returns
+ */
 export async function findPermissionByIdService(
   id,
   selectFields = SELECTED_FIELDS
 ) {
-  if (!id) return null;
   const filter = {};
 
   if (isValidObjectId(id)) {
@@ -63,8 +80,19 @@ export async function findPermissionByIdService(
   return await PermissionModel.findOne(filter).select(selectFields);
 }
 
+/**
+ * Update permission by id
+ * @param {*} id
+ * @param {*} data
+ * @returns
+ */
 export async function updatePermissionByIdService(id, data) {
   const { name } = data;
+  const existPermission = await findPermissionByIdService(id, "_id");
+  if (!existPermission) {
+    throw new NotFoundException("Permission not found");
+  }
+
   const isExistName = await checkExistPermissionNameService(name);
   if (isExistName) {
     throw new BadRequestException("Permission name is exist");
@@ -77,16 +105,32 @@ export async function updatePermissionByIdService(id, data) {
   return permission;
 }
 
+/**
+ * Remove permission by id
+ * @param {*} id
+ * @returns
+ */
 export async function removePermissionByIdService(id) {
+  const existPermission = await findPermissionByIdService(id, "_id");
+  if (!existPermission) {
+    throw new NotFoundException("Permission not found");
+  }
+
   return await PermissionModel.findByIdAndDelete(id).select(
     SELECTED_FIELDS
   );
 }
 
+/**
+ * Check exist permission name
+ * @param {*} name
+ * @param {*} skipId
+ * @returns
+ */
 export async function checkExistPermissionNameService(name, skipId) {
   const existRole = await PermissionModel.findOne({
     name,
     _id: { $ne: skipId },
-  }).select("name");
+  }).select("_id");
   return Boolean(existRole);
 }
